backend-fastapi-tech/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
asyncpg==0.30.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.6.1
